Read CHANGES.md

+++ src/main/java/radon/jujutsu_kaisen/client/gui/overlay/SixEyesOverlay.java
@@ -53,7 +53,7 @@ public class SixEyesOverlay {
                 SorcererUtil.getGrade(client.experience).getName());
         lines.add(gradeText);
 
-        Component experienceText = Component.translatable(String.format("gui.%s.six_eyes_overlay.experience", JujutsuKaisen.MOD_ID), (int) client.experience);
+        Component experienceText = Component.translatable(String.format("gui.%s.six_eyes_overlay.experience", JujutsuKaisen.MOD_ID), client.experience);
         lines.add(experienceText);
 
         int offset = 0;
